+++ CoreLib/IPacket.h
@@ -9,8 +9,8 @@ namespace Core {
     public:
         virtual ~IPacket() = default;
         virtual uint8_t* GetBuffer() = 0;
-        virtual uint16_t GetLength() = 0;
-        virtual void SetLength(uint16_t len) = 0;
+        virtual uint32_t GetLength() = 0;
+        virtual void SetLength(uint32_t len) = 0;
         virtual void Release() = 0;
 
         // Core 에서만 사용
